move FunctionEntryRow into widgets/generic/
