+++ LASSIE/src/widgets/Select.hpp
@@ -7,7 +7,7 @@
 #include <QString>
 
 #include "../core/event_struct.hpp"
-#include "FunctionEntryRow.hpp"
+#include "generic/FunctionEntryRow.hpp"
 
 class FunctionGenerator;
 class EventAttributesViewController;
