+++ LASSIE/src/widgets/Stochos.hpp
@@ -10,7 +10,7 @@
 #include <QString>
 
 #include "../core/event_struct.hpp"
-#include "FunctionEntryRow.hpp"
+#include "generic/FunctionEntryRow.hpp"
 
 class FunctionGenerator;
 class EventAttributesViewController;
