+++ LASSIE/src/widgets/generic/FunctionEntryRow.cpp
