+++ LASSIE/src/widgets/generic/FunctionEntryRow.hpp
